MatchBuilder and Matcher classes
Added MatchBuilder and Matcher classes that enable a match object (Matcher) to be passed around in user code. This object is created through a call to match with no arguments. Creation largely follows the same syntax except for the final function, which must use a || to actually instatiate the match object.